+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/internal/audio/DefaultAudioClientObserver.kt
@@ -351,11 +351,20 @@ class DefaultAudioClientObserver(
         // group by attendee status
         val attendeesByStatus = attendeeUpdates
             .groupBy({ AttendeeStatus.from(it.data) }, { createAttendeeInfo(it) })
+
         attendeesByStatus[AttendeeStatus.Joined]?.let {
             val attendeeJoined = it.minus(currentAttendees)
             if (attendeeJoined.isNotEmpty()) {
                 onAttendeesJoin(attendeeJoined.toTypedArray())
                 currentAttendees.addAll(attendeeJoined)
+            }
+        }
+
+        attendeesByStatus[AttendeeStatus.JoinedNoAudio]?.let {
+            val attendeeJoinedWithoutAudio = it.minus(currentAttendees)
+            if (attendeeJoinedWithoutAudio.isNotEmpty()) {
+                onAttendeesJoinWithoutAudio(attendeeJoinedWithoutAudio.toTypedArray())
+                currentAttendees.addAll(attendeeJoinedWithoutAudio)
             }
         }
 
@@ -386,9 +395,17 @@ class DefaultAudioClientObserver(
         )
     }
 
-    private fun onAttendeesJoin(attendeeInfo: Array<AttendeeInfo>) {
-        logger.debug(TAG, "Added: ${attendeeInfo.joinToString(" ")}")
+    private fun onAttendeesJoinWithoutAudio(attendeeInfo: Array<AttendeeInfo>) {
+        logger.debug(TAG, "Joined without Audio: ${attendeeInfo.joinToString(" ")}")
+        ObserverUtils.notifyObserverOnMainThread(realtimeEventObservers) {
+            it.onAttendeesJoinedWithoutAudio(
+                attendeeInfo
+            )
+        }
+    }
 
+    private fun onAttendeesJoin(attendeeInfo: Array<AttendeeInfo>) {
+        logger.debug(TAG, "Joined: ${attendeeInfo.joinToString(" ")}")
         ObserverUtils.notifyObserverOnMainThread(realtimeEventObservers) {
             it.onAttendeesJoined(
                 attendeeInfo
@@ -398,7 +415,6 @@ class DefaultAudioClientObserver(
 
     private fun onAttendeesLeft(attendeeInfo: Array<AttendeeInfo>) {
         logger.debug(TAG, "Left: ${attendeeInfo.joinToString(" ")}")
-
         ObserverUtils.notifyObserverOnMainThread(realtimeEventObservers) {
             it.onAttendeesLeft(
                 attendeeInfo
@@ -408,7 +424,6 @@ class DefaultAudioClientObserver(
 
     private fun onAttendeesDropped(attendeeInfo: Array<AttendeeInfo>) {
         logger.debug(TAG, "Dropped: ${attendeeInfo.joinToString(" ")}")
-
         ObserverUtils.notifyObserverOnMainThread(realtimeEventObservers) {
             it.onAttendeesDropped(
                 attendeeInfo