+++ app/src/main/java/com/amazonaws/services/chime/sdkdemo/fragment/MeetingFragment.kt
@@ -26,6 +26,7 @@ import android.widget.ImageButton
 import android.widget.LinearLayout
 import android.widget.TextView
 import android.widget.Toast
+import androidx.core.os.bundleOf
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.DiffUtil
@@ -36,10 +37,12 @@ import com.amazonaws.services.chime.sdk.meetings.analytics.EventAttributes
 import com.amazonaws.services.chime.sdk.meetings.analytics.EventName
 import com.amazonaws.services.chime.sdk.meetings.analytics.toJsonString
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.AttendeeInfo
+import com.amazonaws.services.chime.sdk.meetings.audiovideo.AudioVideoConfiguration
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.AudioVideoFacade
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.AudioVideoObserver
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.SignalUpdate
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.VolumeUpdate
+import com.amazonaws.services.chime.sdk.meetings.audiovideo.audio.AudioMode
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.audio.activespeakerdetector.ActiveSpeakerObserver
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.audio.activespeakerpolicy.DefaultActiveSpeakerPolicy
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.contentshare.ContentShareObserver
@@ -58,6 +61,7 @@ import com.amazonaws.services.chime.sdk.meetings.audiovideo.video.gl.EglCoreFact
 import com.amazonaws.services.chime.sdk.meetings.device.DeviceChangeObserver
 import com.amazonaws.services.chime.sdk.meetings.device.MediaDevice
 import com.amazonaws.services.chime.sdk.meetings.device.MediaDeviceType
+import com.amazonaws.services.chime.sdk.meetings.internal.AttendeeStatus
 import com.amazonaws.services.chime.sdk.meetings.realtime.RealtimeObserver
 import com.amazonaws.services.chime.sdk.meetings.realtime.datamessage.DataMessage
 import com.amazonaws.services.chime.sdk.meetings.realtime.datamessage.DataMessageObserver
@@ -147,6 +151,7 @@ class MeetingFragment : Fragment(),
     private lateinit var noVideoOrScreenShareAvailable: TextView
     private lateinit var editTextMessage: EditText
     private lateinit var buttonMute: ImageButton
+    private lateinit var buttonSpeaker: ImageButton
     private lateinit var buttonCamera: ImageButton
     private lateinit var deviceAlertDialogBuilder: AlertDialog.Builder
     private lateinit var additionalOptionsAlertDialogBuilder: AlertDialog.Builder
@@ -169,11 +174,13 @@ class MeetingFragment : Fragment(),
     private lateinit var audioDeviceManager: AudioDeviceManager
 
     companion object {
-        fun newInstance(meetingId: String): MeetingFragment {
+        fun newInstance(meetingId: String, audioVideoConfig: AudioVideoConfiguration): MeetingFragment {
             val fragment = MeetingFragment()
 
-            fragment.arguments =
-                Bundle().apply { putString(HomeActivity.MEETING_ID_KEY, meetingId) }
+            fragment.arguments = bundleOf(
+                HomeActivity.MEETING_ID_KEY to meetingId,
+                HomeActivity.AUDIO_MODE_KEY to audioVideoConfig.audioMode.value
+            )
             return fragment
         }
     }
@@ -234,7 +241,14 @@ class MeetingFragment : Fragment(),
 
         selectTab(meetingModel.tabIndex)
         subscribeToAttendeeChangeHandlers()
-        audioVideo.start()
+        val audioMode = arguments?.getInt(HomeActivity.AUDIO_MODE_KEY)?.let { intValue ->
+            AudioMode.from(intValue, defaultAudioMode = AudioMode.Mono)
+        } ?: AudioMode.Mono
+        val audioVideoConfig = AudioVideoConfiguration(audioMode = audioMode)
+        // Update the Mic & Speaker states
+        updateLocalAttendeeAudioState(audioEnabled = audioVideoConfig.audioMode != AudioMode.NoAudio)
+        // Start Audio Video
+        audioVideo.start(audioVideoConfig)
         audioVideo.startRemoteVideo()
         return view
     }
@@ -243,6 +257,8 @@ class MeetingFragment : Fragment(),
         buttonMute = view.findViewById(R.id.buttonMute)
         buttonMute.setImageResource(if (meetingModel.isMuted) R.drawable.button_mute_on else R.drawable.button_mute)
         buttonMute.setOnClickListener { toggleMute() }
+
+        buttonSpeaker = view.findViewById(R.id.buttonSpeaker)
 
         buttonCamera = view.findViewById(R.id.buttonCamera)
         buttonCamera.setImageResource(if (meetingModel.isCameraOn) R.drawable.button_camera_on else R.drawable.button_camera)
@@ -555,30 +571,11 @@ class MeetingFragment : Fragment(),
     }
 
     override fun onAttendeesJoined(attendeeInfo: Array<AttendeeInfo>) {
-        uiScope.launch {
-            mutex.withLock {
-                attendeeInfo.forEach { (attendeeId, externalUserId) ->
-                    if (DefaultModality(attendeeId).hasModality(ModalityType.Content) &&
-                        !isSelfAttendee(attendeeId) &&
-                        meetingModel.isSharingContent) {
-                        audioVideo.stopContentShare()
-                        screenShareManager?.stop()
-                        val name = meetingModel.currentRoster[DefaultModality(attendeeId).base()]?.attendeeName ?: ""
-                        notifyHandler("$name took over the screen share")
-                    }
-                    meetingModel.currentRoster.getOrPut(
-                        attendeeId,
-                        {
-                            RosterAttendee(
-                                attendeeId,
-                                getAttendeeName(attendeeId, externalUserId)
-                            )
-                        })
-                }
+        onAttendeesJoinedWithStatus(attendeeInfo, AttendeeStatus.Joined)
+    }
 
-                rosterAdapter.notifyDataSetChanged()
-            }
-        }
+    override fun onAttendeesJoinedWithoutAudio(attendeeInfo: Array<AttendeeInfo>) {
+        onAttendeesJoinedWithStatus(attendeeInfo, AttendeeStatus.JoinedNoAudio)
     }
 
     override fun onAttendeesLeft(attendeeInfo: Array<AttendeeInfo>) {
@@ -672,6 +669,48 @@ class MeetingFragment : Fragment(),
             scoresStr,
             LogLevel.DEBUG
         )
+    }
+
+    private fun onAttendeesJoinedWithStatus(attendeeInfo: Array<AttendeeInfo>, status: AttendeeStatus) {
+        uiScope.launch {
+            mutex.withLock {
+                attendeeInfo.forEach { (attendeeId, externalUserId) ->
+                    if (isSelfAttendee(attendeeId))
+                        updateLocalAttendeeAudioState(audioEnabled = status != AttendeeStatus.JoinedNoAudio)
+
+                    if (DefaultModality(attendeeId).hasModality(ModalityType.Content) &&
+                            !isSelfAttendee(attendeeId) &&
+                            meetingModel.isSharingContent) {
+                        audioVideo.stopContentShare()
+                        screenShareManager?.stop()
+                        val name = meetingModel.currentRoster[DefaultModality(attendeeId).base()]?.attendeeName ?: ""
+                        notifyHandler("$name took over the screen share")
+                    }
+                    meetingModel.currentRoster.getOrPut(
+                        attendeeId,
+                        {
+                            RosterAttendee(
+                                attendeeId,
+                                getAttendeeName(attendeeId, externalUserId),
+                                attendeeStatus = status
+                            )
+                        }
+                    )
+                }
+
+                rosterAdapter.notifyDataSetChanged()
+            }
+        }
+    }
+
+    private fun updateLocalAttendeeAudioState(audioEnabled: Boolean) {
+        if (audioEnabled) {
+            buttonMute.visibility = View.VISIBLE
+            buttonSpeaker.visibility = View.VISIBLE
+        } else {
+            buttonMute.visibility = View.GONE
+            buttonSpeaker.visibility = View.GONE
+        }
     }
 
     private fun getAttendeeName(attendeeId: String, externalUserId: String): String {