+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/audiovideo/audio/activespeakerpolicy/DefaultActiveSpeakerPolicy.kt
@@ -11,30 +11,27 @@ import kotlin.math.max
 
 /**
  * [DefaultActiveSpeakerPolicy] A default implementation of the Active Speaker Policy
+ *
+ * @property speakerWeight Specifies how much weight we give to the existing score of the attendee
+ * @property cutoffThreshold Only scores above this threshold will be considered active
+ * @property takeoverRate This constant is used to reduce the active speaker score of other attendees
+ *                        indicating that the current attendee has "possibly" started speaking because
+ *                        the latest volume received was high
+ * @property silenceThreshold The threshold number compared with current received volume.
+ *                            While calculating the new active speaker score, if the current received
+ *                            volume is less than this threshold value, the current received volume is considered as 0,
+ *                            else 1.
  */
-
-class DefaultActiveSpeakerPolicy : ActiveSpeakerPolicy {
+class DefaultActiveSpeakerPolicy @JvmOverloads constructor(
+    private val speakerWeight: Double = 0.9,
+    private val cutoffThreshold: Double = 0.01,
+    private val takeoverRate: Double = 0.2,
+    private val silenceThreshold: Double = 0.2
+) : ActiveSpeakerPolicy {
     /**
      * Map of attendees to their current active speaker scores
      */
     private var speakerScores = mutableMapOf<AttendeeInfo, Double>()
-
-    /**
-     * Specifies how much weight we give to the existing score of the attendee
-     */
-    private val speakerWeight = 0.9
-
-    /**
-     * Only scores above this threshold will be considered active
-     */
-    private val cutoffThreshold = 0.01
-
-    /**
-     * This constant is used to reduce the active speaker score of other attendees
-     * indicating that the current attendee has "possibly" started speaking because
-     * the latest volume received was high
-     */
-    private val takeoverRate = 0.2
 
     /**
      * normalizeFactor is used to convert the VolumeLevel from 1-3 scale to 0-1 scale
@@ -48,13 +45,14 @@ class DefaultActiveSpeakerPolicy : ActiveSpeakerPolicy {
             speakerScores[attendeeInfo] = 0.0
         }
 
-        if (normalizedVolume > 0.0) {
+        if (normalizedVolume > silenceThreshold) {
             normalizedVolume = 1.0
         } else {
             normalizedVolume = 0.0
         }
 
-        val score: Double = (speakerScores.getValue(attendeeInfo) * speakerWeight) + (normalizedVolume * (1.0 - speakerWeight))
+        val score: Double =
+            (speakerScores.getValue(attendeeInfo) * speakerWeight) + (normalizedVolume * (1.0 - speakerWeight))
         speakerScores[attendeeInfo] = score
 
         speakerScores.filterKeys { it != attendeeInfo }.forEach { (otherAttendee, currentScore) ->