+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 ## [Unreleased]
 
+### Fixed
+* [Demo] Fixed demo app crashes when screen share is off and then leave meeting.
+
+### Changed
+* Updated compileSdkVersion and targetSdkVersion from 30 to 31.
+* Changed the silence threshold to 0.2 from 0.0 for `DefaultActiveSpeakerPolicy` (Issue #259) to be more consistent with other platform.
+* Expose weights/rates/thresholds to `DefaultActiveSpeakerPolicy` constructor to make builders easier to customize `DefaultActiveSpeakerPolicy`.
+
+## [0.13.0] - 2021-11-01
+
 ### Added
 * Supports integration with Amazon Transcribe and Amazon Transcribe Medical for live transcription. The Amazon Chime Service uses its active talker algorithm to select the top two active talkers, and sends their audio to Amazon Transcribe (or Amazon Transcribe Medical) in your AWS account. User-attributed transcriptions are then sent directly to every meeting attendee via data messages. Use transcriptions to overlay subtitles, build a transcript, or perform real-time content analysis. For more information, visit [the live transcription guide](https://docs.aws.amazon.com/chime/latest/dg/meeting-transcription.html).